fridge-button

+++ include/models/implementations/instanced/instancedpolygon.hpp
@@ -68,7 +68,7 @@ class InstancedPolygon : public InstancedModel<InstancedPolygonData> {
 		}
 	};
 
-	InstancedPolygon(const UBO &ubo, ScreenParams &screenParams, const vector<Vertex> &vertices, const vector<uint16_t> &indices, shared_ptr<unordered_map<int, InstancedPolygonData>> instances, uint32_t maxInstances = 65536);
+	InstancedPolygon(const UBO &ubo, ScreenParams &screenParams, const vector<Vertex> &vertices, const vector<uint32_t> &indices, shared_ptr<unordered_map<int, InstancedPolygonData>> instances, uint32_t maxInstances = 65536);
 	~InstancedPolygon() override = default;
 
   protected:
